Moved into One Activity - used view model to listen for messages
