+++ app/src/main/AndroidManifest.xml
@@ -13,9 +13,6 @@
         android:theme="@style/Theme.DemoStore"
         tools:targetApi="31">
         <activity
-            android:name=".ui.SunFavsActivity"
-            android:exported="false" />
-        <activity
             android:name=".ui.RecipeActivity"
             android:exported="false" />
         <activity
