+++ app/src/main/java/algonquin/cst2335/demostore/data/SunsetViewModel.java
@@ -7,4 +7,6 @@ import java.util.ArrayList;
 
 public class SunsetViewModel extends ViewModel {
     public MutableLiveData<ArrayList<SunsetData>> dataList = new MutableLiveData<>();
+    public MutableLiveData<SunsetData> selectedSunsetData = new MutableLiveData<>();
+
 }
