+++ app/src/main/java/algonquin/cst2335/demostore/ui/SunActivity.java
@@ -5,15 +5,19 @@ import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import android.content.Context;
-import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
+import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.google.android.material.snackbar.Snackbar;
@@ -24,15 +28,40 @@ import algonquin.cst2335.demostore.R;
 import algonquin.cst2335.demostore.data.SunsetData;
 import algonquin.cst2335.demostore.data.SunsetViewModel;
 import algonquin.cst2335.demostore.databinding.ActivitySunBinding;
-import algonquin.cst2335.demostore.databinding.ActivitySunFavsBinding;
+import algonquin.cst2335.demostore.databinding.SunsetFavItemBinding;
 
 public class SunActivity extends AppCompatActivity {
 
     private static final String TAG = SunActivity.class.getName();
     private ActivitySunBinding binding;
-    private ArrayList<SunsetData> dataList = new ArrayList<>();
-    private SunsetViewModel sunsetViewModel;
-    private boolean hasSearchResult;
+    protected ArrayList<SunsetData> dataList = new ArrayList<>();
+    protected SunsetViewModel sunsetViewModel;
+    private RecyclerView.Adapter<MyRowHolder> myAdapter;
+    private int deletePosition;
+    private View posView;
+    private boolean canDelete = false;
+    class MyRowHolder extends RecyclerView.ViewHolder {
+        TextView pos;
+        TextView sunrise;
+        TextView sunset;
+        TextView dawn;
+        TextView dusk;
+
+        public MyRowHolder(@NonNull View itemView) {
+            super(itemView);
+            pos = itemView.findViewById(R.id.sunsetPosSunriseValue);
+            sunrise = itemView.findViewById(R.id.sunsetFavSunriseValue);
+            sunset = itemView.findViewById(R.id.sunsetFavSunsetValue);
+            dawn = itemView.findViewById(R.id.sunsetFavDawnValue);
+            dusk = itemView.findViewById(R.id.sunsetFavDuskValue);
+
+            itemView.setOnClickListener(view -> {
+                deletePosition = getAdapterPosition();
+                posView = pos;
+                canDelete = true;
+            });
+        }
+    }
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -52,7 +81,6 @@ public class SunActivity extends AppCompatActivity {
 
         latField.setText(latData);
         longField.setText(longData);
-        hasSearchResult = false;
 
         sunsetViewModel = new ViewModelProvider(this).get(SunsetViewModel.class);
         dataList = sunsetViewModel.dataList.getValue();
@@ -60,6 +88,63 @@ public class SunActivity extends AppCompatActivity {
         if (dataList == null) {
             sunsetViewModel.dataList.postValue(dataList = new ArrayList<SunsetData>());
         }
+
+        sunsetViewModel = new ViewModelProvider(this).get(SunsetViewModel.class);
+        binding.sunsetFavsRecyclerView.setLayoutManager(new LinearLayoutManager(this));
+        dataList = sunsetViewModel.dataList.getValue();
+
+        if (dataList == null) {
+            sunsetViewModel.dataList.postValue(dataList = new ArrayList<SunsetData>());
+        }
+
+        binding.sunsetFavsRecyclerView.setAdapter(myAdapter = new RecyclerView.Adapter<MyRowHolder>() {
+            @NonNull
+            @Override
+            public MyRowHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
+                SunsetFavItemBinding svBinding = SunsetFavItemBinding.inflate(getLayoutInflater());
+                return new MyRowHolder(svBinding.getRoot());
+            }
+
+            @Override
+            public void onBindViewHolder(@NonNull MyRowHolder holder, int position) {
+                SunsetData sunsetData = dataList.get(position);
+                String posText = String.valueOf(position + 1);
+
+                holder.pos.setText(posText);
+                holder.sunrise.setText(sunsetData.getSunrise());
+                holder.sunset.setText(sunsetData.getSunset());
+                holder.dawn.setText(sunsetData.getDawn());
+                holder.dusk.setText(sunsetData.getDusk());
+            }
+
+            @Override
+            public int getItemCount() {
+                return dataList.size();
+            }
+        });
+
+        // add test data
+        SunsetData sunsetDataTest1 = new SunsetData(
+                "6:00 AM",
+                "7:30 PM",
+                "5:15 AM",
+                "8:15 PM"
+        );
+        SunsetData sunsetDataTest2 = new SunsetData(
+                "7:15 AM",
+                "8:00 PM",
+                "6:45 AM",
+                "8:30 PM"
+        );
+        dataList.add(sunsetDataTest1);
+        myAdapter.notifyItemInserted(dataList.size() - 1);
+        dataList.add(sunsetDataTest2);
+        sunsetViewModel.dataList.postValue(dataList);
+        myAdapter.notifyItemInserted(dataList.size() - 1);
+
+        sunsetViewModel.dataList.observe(this, (list) -> {
+         Log.d(TAG, "ADDED MSG" + list.toString());
+        });
 
         searchBtn.setOnClickListener(click -> {
             SharedPreferences.Editor editor = preferences.edit();
@@ -70,10 +155,16 @@ public class SunActivity extends AppCompatActivity {
             editor.putString("Longitude", longText);
             editor.apply();
 
-            hasSearchResult = true;
-
             if (!longText.isEmpty() && !latText.isEmpty()) {
-                // do stuff...
+                Log.d(TAG, "testing. 1");
+
+                SunDetailsFragment detailsFragment = new SunDetailsFragment(dataList.get(0));
+
+                getSupportFragmentManager()
+                        .beginTransaction()
+                        .replace(R.id.fragmentLocation, detailsFragment)
+                        .addToBackStack("")
+                        .commit();
             }
         });
     }
@@ -93,17 +184,40 @@ public class SunActivity extends AppCompatActivity {
             // display help dialog
             AlertDialog.Builder builder = new AlertDialog.Builder(SunActivity.this);
             builder.setMessage(R.string.sunset_help_main)
-                    .setTitle(R.string.sunset_help)
-                    .setPositiveButton(R.string.sunset_ok, (dialogInterface, i) -> {})
-                    .create().show();
-        } else if (choice == R.id.sunsetFavs) {
-            Intent intent = new Intent(SunActivity.this, SunFavsActivity.class);
-            startActivity(intent);
-        } else if (choice == R.id.sunsetMakeFav) {
-            if (hasSearchResult) {
-                Toast.makeText(this, R.string.sunset_saved_to_favs, Toast.LENGTH_LONG).show();
+                .setTitle(R.string.sunset_help)
+                .setPositiveButton(R.string.sunset_ok, (dialogInterface, i) -> {})
+                .create().show();
+        } else if (choice == R.id.sunsetFavsDelete) {
+            if (!canDelete) {
+                Toast.makeText(this, R.string.sunset_nothing_to_del, Toast.LENGTH_SHORT).show();
             } else {
-                Toast.makeText(this, R.string.sunset_nothing, Toast.LENGTH_LONG).show();
+                AlertDialog.Builder builder = new AlertDialog.Builder(this);
+                String msgText = getResources().getString(R.string.sunset_del_fav_msg);
+                String undoTextPos = String.valueOf(deletePosition + 1);
+
+                builder.setMessage(msgText + undoTextPos)
+                    .setTitle(R.string.sunset_del_fav_title)
+                    .setNegativeButton(R.string.sunset_no, (dialogInterface, i) -> {})
+                    .setPositiveButton(R.string.sunset_yes, (dialogInterface, i) -> {
+                        SunsetData data = dataList.get(deletePosition);
+                        String undoText = getResources().getString(R.string.sunset_del_succ);
+
+                        dataList.remove(deletePosition);
+                        myAdapter.notifyItemRemoved(deletePosition);
+
+                        Snackbar.make(
+                            posView,
+                            undoText + undoTextPos,
+                            Snackbar.LENGTH_LONG)
+                        .setAction(R.string.sunset_undo, clk -> {
+                            dataList.add(deletePosition, data);
+                            myAdapter.notifyItemInserted(deletePosition);
+                        })
+                        .show();
+
+                        canDelete = false;
+                    })
+                    .create().show();
             }
         }
         return true;
