--- app/src/main/res/layout/activity_sun_favs.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.SunFavsActivity">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/sunsetFavsToolbar"
-        android:layout_width="409dp"
-        android:layout_height="wrap_content"
-        android:background="?attr/colorPrimary"
-        android:minHeight="?attr/actionBarSize"
-        android:theme="?attr/actionBarTheme"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:title="@string/sunset_app_name"
-        app:titleTextColor="#E8CEAE" />
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/sunsetFavsRecyclerView"
-        android:layout_width="409dp"
-        android:layout_height="665dp"
-        android:layout_marginTop="64dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/sunsetFavsToolbar" />
-</androidx.constraintlayout.widget.ConstraintLayout>
