--- app/src/main/res/menu/sunset_favs_menu.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:android="http://schemas.android.com/apk/res/android">
-
-    <item
-        android:id="@+id/sunsetFavsDelete"
-        android:title="@string/sunset_delete"
-        android:icon="@drawable/ic_sunset_delete"
-        app:showAsAction="always"
-    />
-    <item
-        android:id="@+id/sunsetFavsHelp"
-        android:title="@string/sunset_help"
-        app:showAsAction="never"
-    />
-    <item
-        android:id="@+id/sunsetFavsSearch"
-        android:title="@string/sunset_search_activity"
-        app:showAsAction="never"
-    />
-</menu>
